Move two more functions to nodes.py
